match making... working on!

+++ main.coffee
@@ -10,16 +10,22 @@ opts =
 server = ether(plugins,opts)
 
 article_test = (col,root) ->
-	server.publishDocs 'test', (client,cb) -> col.findAll({},cb)	
-	root.addItem = (client,cb) -> 
-		col.save {hello:'world'}, cb
-	root.deleteItem = (client,id,cb) -> 
-		col.remove {_id:db.ObjectId(id)},cb
+	server.publishDocs 'test', (client,cb) -> col.find({}).limit(10).sort when:-1,cb
+	root.addItem = (client,msg,cb) -> col.save {author:client.tag(),msg:msg,when:Date.now()}, cb
+	root.deleteItem = (client,id,cb) -> col.remove {_id:db.ObjectId(id),author:client.tag()},cb
 	root.likeItem = (client,id,cb) -> 
-		col.update {_id:db.ObjectId(id)},$inc:like:1, cb
+		q = 
+			_id:db.ObjectId(id)			
+			liked:$not:$elemMatch:client.tag()
+		u = 
+			$push:liked:client.tag()
+			$inc:like:1
+
+		col.update q, u, db.expect('failed',cb,1)
 
 server?.use 'room'
 server?.use 'rsvp'
+server?.use 'matchmaking'
 
 if server
 	{db,rpc,deps} = server
